# Performance improvement and reduction in memory allocation when dealing with Formatted cells