--- a/ClosedXML.Tests.Performance/BenchmarkTestRunner.cs
+++ b/ClosedXML.Tests.Performance/BenchmarkTestRunner.cs
@@ -1,0 +1,11 @@
+namespace ClosedXML.Tests.Performance;
+
+[TestFixture]
+public class BenchmarkTestRunner
+{
+    [Test, Explicit("Performance test")]
+    public void WorkbookOperationPerformanceTests()
+    {
+        BenchmarkRunnerHelper.RunBenchmark<WorkbookOperationBenchmarks>();
+    }
+}