… ClosedXML/Excel/Caching/XLRepositoryBase.cs
… ClosedXML/Excel/Cells/XLCell.cs
… ClosedXML/Excel/Style/XLStyleKey.cs
… ClosedXML/Properties/AssemblyInfo.cs
… ClosedXML/Utils/CollectionPools.cs